backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
SQLAlchemy==1.4.23
psycopg2-binary==2.9.1
fastapi-socketio==0.0.10
python-dotenv==0.19.0
Flask==2.0.1
Flask-SocketIO==5.1.1
Werkzeug==2.0.1
